[Fix] Todo 제거 기능 및 check 기능 함수 위치 수정
